requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==4.2.1
django-cloudinary-storage==0.3.0
django-pdb==0.6.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
